pyproject.toml
[project]
name = "ts-funnel"
version = "0.1.0"
description = "Auto-discover and funnel local dev servers through Tailscale"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "tomli>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ts-funnel = "ts_funnel.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
